Bug fixes, file restructure
